#32 feat: 공유하기 수정
