+++ components/letterList/index.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useEffect } from 'react';
 import { useRouter } from 'next/router';
 
 // slick
@@ -9,13 +9,17 @@ import 'slick-carousel/slick/slick-theme.css';
 // style
 import { DiarySingleLetterListWrapper } from './styled';
 
-// kakao
-import useSNSShare from './useSNSShare';
-
 interface IProps {
   petInfo: any;
   type: string;
   name: string;
+}
+
+// 카카오 SDK는 window 객체에 전역으로 추가되니 TS를 위해 선언
+declare global {
+  interface Window {
+    Kakao: any;
+  }
 }
 
 const DiarySingleLetterList = ({ petInfo, type, name }: IProps) => {
@@ -26,6 +30,47 @@ const DiarySingleLetterList = ({ petInfo, type, name }: IProps) => {
     infinite: false,
     speed: 500,
   };
+
+  // 카카오톡 공유하기
+  const shareToKakaoTalk = (userId: any) => {
+    // @ts-ignore
+    if (window.Kakao === undefined) {
+      return;
+    }
+  
+    const kakao = window.Kakao;
+  
+    // 인증이 안되어 있는 경우, 인증한다.
+    if (!kakao.isInitialized()) {
+      kakao.init('19ef6355b4d2cb8afc9c439c165794be');
+      console.log(kakao.isInitialized());
+    }
+  
+    // 메시지 템플릿 활용
+      window.Kakao.Link.createCustomButton({
+        container: '#kakao-link-btn',
+        templateId: 90948,
+        templateArgs: {
+          userId: `${userId}`,
+        },
+      });
+    };
+    const onShareKakaoClick = () => {
+      shareToKakaoTalk(userId);
+    };
+
+    // 카카오 sdk 선언
+      useEffect(() => {
+        const script = document.createElement('script');
+        script.src = 'https://developers.kakao.com/sdk/js/kakao.js';
+        script.async = true;
+
+        document.body.appendChild(script);
+
+        return () => {
+        document.body.removeChild(script);
+        };
+      }, []);
 
   return (
     <DiarySingleLetterListWrapper>
@@ -61,7 +106,11 @@ const DiarySingleLetterList = ({ petInfo, type, name }: IProps) => {
               : ''}
           </Slider>
           <div className="ios-share-img">
-            <img src="/images/ios-share.png" alt="카카오로 공유하기" onClick={useSNSShare} />
+            <img 
+              id="kakao-link-btn" 
+              src="/images/ios-share.png" 
+              alt="카카오톡 공유하기"
+              onClick={onShareKakaoClick}/>
           </div>
           <div className="bottom-btn" onClick={() => router.push('/diary')}>
             <span>곰곰이 일기 다시보기</span>
